Redirect to HTTPS

+++ wsgi/PointTracker/pointtracker/views.py
@@ -17,7 +17,7 @@ from PointTracker import Return_Reward_Program
 from PointTracker import Remove_PT_account_Reward_Program_Passwords
 
 #from PointTracker import hack_mongo
-
+import Globalvars
 
 
 
@@ -25,6 +25,9 @@ from PointTracker import Remove_PT_account_Reward_Program_Passwords
 
 @view_config(route_name='home', renderer='pointtracker:templates/index.html')               #pull up our index.html
 def server_view0(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                       #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     if 'PointTracker_Login' in request.cookies:                                             #is the cookie set?
         url = request.route_url('pointtracker')                                             #cookie was set so redirect
         return HTTPFound(location=url)
@@ -34,6 +37,9 @@ def server_view0(request):
 
 @view_config(renderer="json", name="Sign_In_View")
 def server_view1(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     hasAccount = request.hasPTaccount()
     if hasAccount:                                                          #true?
         request.remember_me()                                               #remember me will determine cookie time limit
@@ -43,6 +49,9 @@ def server_view1(request):
 
 @view_config(renderer="json", name="Register_View")
 def server_view2(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
 #    hack_mongo(request.POST)
     hasAccount = request.hasPTaccount()                                     #the account already exists so don't create a new one and warn user
     if not hasAccount:
@@ -54,6 +63,9 @@ def server_view2(request):
 
 @view_config(route_name='Sign_Out', renderer='pointtracker:templates/index.html')               #pull up our index.html
 def server_view3(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     request.forget()                                                                            #delete cookie
     return {}
 
@@ -61,6 +73,9 @@ def server_view3(request):
 
 @view_config(route_name='pointtracker', renderer='pointtracker:templates/pointtracker.html')               #pull up our index.html
 def server_view4(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     if 'PointTracker_Login' not in request.cookies:                             #is the cookie set?
         url = request.route_url('home')                                     #cookie is not set so redirect home sign in page
         return HTTPFound(location=url)
@@ -70,6 +85,9 @@ def server_view4(request):
 
 @view_config(renderer="json", name="Get_PointTracker_Account_View")        #Get our PT_account of of the database
 def server_view5(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     _id = request.cookies['PointTracker_Login']                            # cookie is in here and contains the _id
     PT_account = Get_PointTracker_Account(_id)                             # Pull the PT_account out of the database
     Remove_PT_account_Reward_Program_Passwords(PT_account)                 # we need to remove RP_account passwords so that they can't be seen on client side
@@ -79,6 +97,9 @@ def server_view5(request):
 
 @view_config(renderer="json", name="Update_PointTracker_Account_View")
 def server_view6(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     RP_account = Refresh_Reward_Program(request.GET)                       #Get the new updated Reward Program
     return RP_account
 
@@ -86,6 +107,9 @@ def server_view6(request):
 
 @view_config(renderer="json", name="Add_Sub_Account_View")
 def server_view7(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     Add_Sub_Account(request.GET)                                            #Add a new sub account to the database
     return
 
@@ -93,6 +117,9 @@ def server_view7(request):
 
 @view_config(renderer="json", name="Delete_Sub_Account_View")
 def server_view8(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     Delete_Sub_Account(request.GET)                                         #Delete a sub account out of the database
     return
 
@@ -100,6 +127,9 @@ def server_view8(request):
 
 @view_config(renderer="json", name="Add_Reward_Program_View")
 def server_view9(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     RP_account = Add_Reward_Program(request.GET)                            #Add a Reward Program to the database
     return RP_account                                                       #Check later if needed for callbacks
 
@@ -107,6 +137,9 @@ def server_view9(request):
 
 @view_config(renderer="json", name="Delete_Reward_Program_View")
 def server_view10(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     Delete_Reward_Program(request.GET)                                      #Delete a Reward Program from the database
     return
 
@@ -114,6 +147,9 @@ def server_view10(request):
 
 @view_config(renderer="json", name="Refresh_Reward_Program_View")
 def server_view11(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     RP_account = Refresh_Reward_Program(request.GET)                        #Refreshes by scraping and updating the database
     return RP_account                                                       #Return for callback info inside
 
@@ -121,6 +157,9 @@ def server_view11(request):
 
 @view_config(renderer="json", name="Edit_Reward_Program_View")
 def server_view12(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     RP_account = Edit_Reward_Program(request.GET)                           #Edit/Modify a Reward Program in the database
     return RP_account                                                       #Check later if needed for callbacks
 
@@ -128,6 +167,9 @@ def server_view12(request):
 
 @view_config(renderer="json", name="Get_Reward_Program_View")
 def server_view13(request):
+    if request.url.startswith('http:') and Globalvars.DEPLOY == True:                   #Only redirect if DEPLOY if on live website
+        return HTTPFound('https' + request.url[4:])                                         #Redirect traffic to https
+
     RP_account = Return_Reward_Program(request.GET)                         #Returns Reward Program from database  (used for testing purposes without scraping and updating database
     RP_account['RP_password'] = ''                                          #Remove password for client side for security reasons. no reason for the client to have it.
     return RP_account
